refactor(viewer): split createViewerStore into domain slices
Break the ~700-line createViewerStore into per-domain zustand slices —
slices/viewer.{core,view,channels,overlays}.store.ts — composed via the
shared ViewerSlice type (alongside the annotations slice from the prior
commit). createViewerStore is now a thin composition root (id + slice
spreads + persist/migrate config); ViewerStoreState is just `id` and
ViewerStoreActions is gone — each domain owns its state + actions.

Pure structural move: store shape identical, 1407 unit tests green,
0 type errors.
